separate schema out
